IntegerField in models

+++ app/models.py
@@ -300,7 +300,7 @@ class Notice(models.Model):
 #------------------------------------------------------------------------------
 class Inventory_history(models.Model):
     material = models.ForeignKey(Material, on_delete=models.CASCADE,verbose_name = " قطعه ")
-    quantity = models.DecimalField(default='1', max_digits=30, decimal_places=4, verbose_name = " تعداد ")
+    quantity = models.IntegerField(default='1', verbose_name = " تعداد ")
     manager = models.ForeignKey(User, on_delete=models.CASCADE,verbose_name = "مسئول")
     time = models.DateTimeField(auto_now_add=True, verbose_name = "زمان")
     station = models.ForeignKey(Station, on_delete=models.CASCADE, verbose_name = " ایستگاه ")
@@ -322,7 +322,7 @@ class Inventory_history(models.Model):
 #------------------------------------------------------------------------------
 class Station_exit_history(models.Model):
     material = models.ForeignKey(Material, on_delete=models.CASCADE,verbose_name = " قطعه ")
-    quantity = models.DecimalField(default='1', max_digits=30, decimal_places=4, verbose_name = " تعداد ")
+    quantity = models.IntegerField(default='1', verbose_name = " تعداد ")
     manager = models.ForeignKey(User, on_delete=models.CASCADE,verbose_name = "مسئول")
     time = models.DateTimeField(auto_now_add=True, verbose_name = "زمان")
     station = models.ForeignKey(Station, on_delete=models.CASCADE, verbose_name = " ایستگاه ")
